pyproject.toml
[tool.poetry]
name = "contrapi"
version = "0.1.0"
description = "api contraloria"
authors = ["ticbox"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.88.0"}
sqlalchemy = "^1.4.45"
pymysql = "^1.0.2"
sqlalchemy-utils = "^0.38.3"
alembic = "^1.9.0"
configparser = "^5.3.0"
mysql-connector = "^2.2.9"
mysql-connector-python = "^8.0.31"
mysqlclient = "^2.1.1"
pyjwt = "^2.6.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
